+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/attributes/service_attributes.py
@@ -17,7 +17,7 @@ from typing import Final
 SERVICE_NAME: Final = "service.name"
 """
 Logical name of the service.
-Note: MUST be the same for all instances of horizontally scaled services. If the value was not specified, SDKs MUST fallback to `unknown_service:` concatenated with [`process.executable.name`](process.md#process), e.g. `unknown_service:bash`. If `process.executable.name` is not available, the value MUST be set to `unknown_service`.
+Note: MUST be the same for all instances of horizontally scaled services. If the value was not specified, SDKs MUST fallback to `unknown_service:` concatenated with [`process.executable.name`](process.md), e.g. `unknown_service:bash`. If `process.executable.name` is not available, the value MUST be set to `unknown_service`.
 """
 
 SERVICE_VERSION: Final = "service.version"